Split API mock and client test running into separate packages
