+++ client/service_offering_test.go
@@ -15,7 +15,7 @@ func TestServiceOfferings(t *testing.T) {
 	tests := []RouteTest{
 		{
 			Description: "Delete service offering",
-			Route: MockRoute{
+			Route: testutil.MockRoute{
 				Method:   "DELETE",
 				Endpoint: "/v3/service_offerings/928a32d9-8101-4b86-85a4-96e06f833c2d",
 				Status:   http.StatusNoContent,
@@ -27,7 +27,7 @@ func TestServiceOfferings(t *testing.T) {
 		},
 		{
 			Description: "Get service offering",
-			Route: MockRoute{
+			Route: testutil.MockRoute{
 				Method:   "GET",
 				Endpoint: "/v3/service_offerings/928a32d9-8101-4b86-85a4-96e06f833c2d",
 				Output:   []string{so},
@@ -39,7 +39,7 @@ func TestServiceOfferings(t *testing.T) {
 		},
 		{
 			Description: "List all service offerings",
-			Route: MockRoute{
+			Route: testutil.MockRoute{
 				Method:   "GET",
 				Endpoint: "/v3/service_offerings",
 				Output:   g.Paged([]string{so}, []string{so2}),
@@ -51,7 +51,7 @@ func TestServiceOfferings(t *testing.T) {
 		},
 		{
 			Description: "Update service offering",
-			Route: MockRoute{
+			Route: testutil.MockRoute{
 				Method:   "PATCH",
 				Endpoint: "/v3/service_offerings/928a32d9-8101-4b86-85a4-96e06f833c2d",
 				Output:   []string{so},
